Test Inheritance works properly for both types.
